+++ cms/lib/Api/Router.php
@@ -61,9 +61,6 @@ class Router
                 case 'theme':
                     ThemeEditorController::handle($rest, $method, $config);
                     return;
-                case 'blocks':
-                    BlocksController::handle($rest, $method, $config);
-                    return;
                 case 'backup':
                     BackupController::handle($rest, $method, $config);
                     return;